refactor(AudioUploader): move on AudioUploader to adapters/uploaders/
