+++ app/adapters/__init__.py
@@ -1,1 +1,2 @@
 from .senders import EmailSenderAdapter
+from .uploaders import AudioAwsUploader
