+++ app/adapters/uploaders/__init__.py
@@ -1,0 +1,1 @@
+from .aws import AudioAwsUploader
