+++ app/services/__init__.py
@@ -1,2 +1,2 @@
-from .audio import AudioConverter, AudioFile, AudioRecognizer, AudioUploader
+from .audio import AudioConverter, AudioFile, AudioRecognizer
 from .uploader import AwsUploader
